+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/SubprojectDao.java
@@ -2,11 +2,12 @@ package uk.ac.bbsrc.tgac.miso.persistence;
 
 import java.util.List;
 
+import uk.ac.bbsrc.tgac.miso.core.data.Project;
 import uk.ac.bbsrc.tgac.miso.core.data.Subproject;
 
 public interface SubprojectDao extends SaveDao<Subproject> {
 
-  public Subproject getByAlias(String alias);
+  public Subproject getByProjectAndAlias(Project project, String alias);
 
   public List<Subproject> listByProjectId(Long projectId);