+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateSubprojectDao.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import uk.ac.bbsrc.tgac.miso.core.data.Project;
 import uk.ac.bbsrc.tgac.miso.core.data.Subproject;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.DetailedSampleImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SubprojectImpl;
@@ -85,8 +86,9 @@ public class HibernateSubprojectDao implements SubprojectDao {
   }
 
   @Override
-  public Subproject getByAlias(String alias) {
+  public Subproject getByProjectAndAlias(Project project, String alias) {
     return (Subproject) currentSession().createCriteria(SubprojectImpl.class)
+        .add(Restrictions.eq("parentProject", project))
         .add(Restrictions.eq("alias", alias))
         .uniqueResult();
   }